+++ service-api/src/main/java/greencity/service/ubs/UBSManagementService.java
@@ -294,24 +294,6 @@ public interface UBSManagementService {
     void updateUserViolation(AddingViolationsToUserDto add, MultipartFile[] multipartFiles);
 
     /**
-     * Method that return parameters for building table on admin's page.
-     *
-     * @param userId of {@link Long} administrator's user id;
-     * @author Liubomyr Pater
-     */
-    TableParamsDTO getParametersForOrdersTable(Long userId);
-
-    /**
-     * Method that return.
-     *
-     * @param userUuid                     of {@link String}
-     * @param requestToChangeOrdersDataDTO of {@link RequestToChangeOrdersDataDTO}
-     * @author Liubomyr Pater
-     */
-    ChangeOrderResponseDTO chooseOrdersDataSwitcher(String userUuid,
-        RequestToChangeOrdersDataDTO requestToChangeOrdersDataDTO);
-
-    /**
      * Method that save ReasonNotTakeBagDto.
      */
     ReasonNotTakeBagDto saveReason(Long orderId, String description, List<MultipartFile> images);
